feat(DTFS2-7395): move mock gef facilities to mock fixtures
