+++ e2e-tests/gef/cypress/e2e/clone/clone-gef-deal.spec.js
@@ -12,7 +12,7 @@ import mandatoryCriteria from '../pages/mandatory-criteria';
 import uploadFiles from '../pages/upload-files';
 import statusBanner from '../pages/application-status-banner';
 import CONSTANTS from '../../fixtures/constants';
-import { anUnissuedCashFacility } from '../../fixtures/mocks/mock-facilities';
+import { anUnissuedCashFacility } from '../../../../e2e-fixtures/mock-gef-facilities';
 import { BANK1_MAKER1, BANK1_CHECKER1 } from '../../../../e2e-fixtures/portal-users.fixture';
 import { MOCK_APPLICATION_MIN } from '../../fixtures/mocks/mock-deals';
