+++ e2e-tests/gef/cypress/e2e/facility/unissued-facilities/AIN/unissued-facility-to-issued-cover-start-date-AIN.spec.js
@@ -8,7 +8,7 @@ import dateConstants from '../../../../../../e2e-fixtures/dateConstants';
 
 import { MOCK_APPLICATION_AIN } from '../../../../fixtures/mocks/mock-deals';
 import { BANK1_MAKER1 } from '../../../../../../e2e-fixtures/portal-users.fixture';
-import { anUnissuedCashFacility } from '../../../../fixtures/mocks/mock-facilities';
+import { anUnissuedCashFacility } from '../../../../../../e2e-fixtures/mock-gef-facilities';
 
 import { mainHeading } from '../../../partials';
 import applicationPreview from '../../../pages/application-preview';
